refactor(cli): split events.rs into per-concern submodules (#179)
Promotes events.rs (2,101 lines) → events/mod.rs and splits by concern:
- types.rs: EventType enum, Event struct
- filter.rs: EventFilter, EventSubscription
- broadcaster.rs: EventBroadcaster impl
- listener.rs: EventListener struct + impl

Public API is unchanged; mod.rs re-exports all types.
Part of the context-burn reduction refactor series.
