+++ apps/things3-cli/src/events/mod.rs
@@ -1,414 +1,23 @@
 //! Event broadcasting system for task/project changes
 
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::sync::Arc;
-use things3_core::{Result, ThingsId};
-use tokio::sync::{broadcast, RwLock};
-use uuid::Uuid;
+mod broadcaster;
+mod filter;
+mod listener;
+mod types;
 
-use crate::progress::ProgressUpdate;
-
-/// Event types for Things 3 entities
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(tag = "event_type")]
-pub enum EventType {
-    /// Task events
-    TaskCreated {
-        task_id: ThingsId,
-    },
-    TaskUpdated {
-        task_id: ThingsId,
-    },
-    TaskDeleted {
-        task_id: ThingsId,
-    },
-    TaskCompleted {
-        task_id: ThingsId,
-    },
-    TaskCancelled {
-        task_id: ThingsId,
-    },
-
-    /// Project events
-    ProjectCreated {
-        project_id: ThingsId,
-    },
-    ProjectUpdated {
-        project_id: ThingsId,
-    },
-    ProjectDeleted {
-        project_id: ThingsId,
-    },
-    ProjectCompleted {
-        project_id: ThingsId,
-    },
-
-    /// Area events
-    AreaCreated {
-        area_id: ThingsId,
-    },
-    AreaUpdated {
-        area_id: ThingsId,
-    },
-    AreaDeleted {
-        area_id: ThingsId,
-    },
-
-    /// Progress events
-    ProgressStarted {
-        operation_id: Uuid,
-    },
-    ProgressUpdated {
-        operation_id: Uuid,
-    },
-    ProgressCompleted {
-        operation_id: Uuid,
-    },
-    ProgressFailed {
-        operation_id: Uuid,
-    },
-}
-
-/// Event data structure
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct Event {
-    pub id: Uuid,
-    pub event_type: EventType,
-    pub timestamp: DateTime<Utc>,
-    pub data: Option<serde_json::Value>,
-    pub source: String,
-}
-
-/// Event filter for subscriptions
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct EventFilter {
-    pub event_types: Option<Vec<EventType>>,
-    pub entity_ids: Option<Vec<ThingsId>>,
-    pub sources: Option<Vec<String>>,
-    pub since: Option<DateTime<Utc>>,
-}
-
-impl EventFilter {
-    /// Check if an event matches this filter
-    #[must_use]
-    pub fn matches(&self, event: &Event) -> bool {
-        // Check event types
-        if let Some(ref types) = self.event_types {
-            if !types
-                .iter()
-                .any(|t| std::mem::discriminant(t) == std::mem::discriminant(&event.event_type))
-            {
-                return false;
-            }
-        }
-
-        // Check entity IDs (applies to task/project/area events; progress events have no entity ID)
-        if let Some(ref ids) = self.entity_ids {
-            let event_entity_id: Option<&ThingsId> = match &event.event_type {
-                EventType::TaskCreated { task_id }
-                | EventType::TaskUpdated { task_id }
-                | EventType::TaskDeleted { task_id }
-                | EventType::TaskCompleted { task_id }
-                | EventType::TaskCancelled { task_id } => Some(task_id),
-                EventType::ProjectCreated { project_id }
-                | EventType::ProjectUpdated { project_id }
-                | EventType::ProjectDeleted { project_id }
-                | EventType::ProjectCompleted { project_id } => Some(project_id),
-                EventType::AreaCreated { area_id }
-                | EventType::AreaUpdated { area_id }
-                | EventType::AreaDeleted { area_id } => Some(area_id),
-                EventType::ProgressStarted { .. }
-                | EventType::ProgressUpdated { .. }
-                | EventType::ProgressCompleted { .. }
-                | EventType::ProgressFailed { .. } => None,
-            };
-
-            if let Some(entity_id) = event_entity_id {
-                if !ids.contains(entity_id) {
-                    return false;
-                }
-            }
-        }
-
-        // Check sources
-        if let Some(ref sources) = self.sources {
-            if !sources.contains(&event.source) {
-                return false;
-            }
-        }
-
-        // Check timestamp
-        if let Some(since) = self.since {
-            if event.timestamp < since {
-                return false;
-            }
-        }
-
-        true
-    }
-}
-
-/// Event subscription
-#[derive(Debug, Clone)]
-pub struct EventSubscription {
-    pub id: Uuid,
-    pub filter: EventFilter,
-    pub sender: broadcast::Sender<Event>,
-}
-
-/// Event broadcaster for managing and broadcasting events
-pub struct EventBroadcaster {
-    sender: broadcast::Sender<Event>,
-    subscriptions: Arc<RwLock<HashMap<Uuid, EventSubscription>>>,
-}
-
-impl EventBroadcaster {
-    /// Create a new event broadcaster
-    #[must_use]
-    pub fn new() -> Self {
-        let (sender, _) = broadcast::channel(1000);
-        Self {
-            sender,
-            subscriptions: Arc::new(RwLock::new(HashMap::new())),
-        }
-    }
-
-    /// Subscribe to events with a filter
-    pub async fn subscribe(&self, filter: EventFilter) -> broadcast::Receiver<Event> {
-        let subscription_id = Uuid::new_v4();
-        let (sub_sender, receiver) = broadcast::channel(100);
-
-        let subscription = EventSubscription {
-            id: subscription_id,
-            filter,
-            sender: sub_sender,
-        };
-
-        {
-            let mut subscriptions = self.subscriptions.write().await;
-            subscriptions.insert(subscription_id, subscription);
-        }
-
-        receiver
-    }
-
-    /// Unsubscribe from events
-    pub async fn unsubscribe(&self, subscription_id: Uuid) {
-        let mut subscriptions = self.subscriptions.write().await;
-        subscriptions.remove(&subscription_id);
-    }
-
-    /// Broadcast an event
-    ///
-    /// # Errors
-    /// Returns an error if broadcasting fails
-    pub async fn broadcast(&self, event: Event) -> Result<()> {
-        // Send to main channel (ignore if no receivers)
-        let _ = self.sender.send(event.clone());
-
-        // Send to filtered subscriptions
-        let subscriptions = self.subscriptions.read().await;
-        for subscription in subscriptions.values() {
-            if subscription.filter.matches(&event) {
-                let _ = subscription.sender.send(event.clone());
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Create and broadcast a task event
-    ///
-    /// # Errors
-    /// Returns an error if broadcasting fails
-    pub async fn broadcast_task_event(
-        &self,
-        event_type: EventType,
-        _task_id: ThingsId,
-        data: Option<serde_json::Value>,
-        source: &str,
-    ) -> Result<()> {
-        let event = Event {
-            id: Uuid::new_v4(),
-            event_type,
-            timestamp: Utc::now(),
-            data,
-            source: source.to_string(),
-        };
-
-        self.broadcast(event).await
-    }
-
-    /// Create and broadcast a project event
-    ///
-    /// # Errors
-    /// Returns an error if broadcasting fails
-    pub async fn broadcast_project_event(
-        &self,
-        event_type: EventType,
-        _project_id: ThingsId,
-        data: Option<serde_json::Value>,
-        source: &str,
-    ) -> Result<()> {
-        let event = Event {
-            id: Uuid::new_v4(),
-            event_type,
-            timestamp: Utc::now(),
-            data,
-            source: source.to_string(),
-        };
-
-        self.broadcast(event).await
-    }
-
-    /// Create and broadcast an area event
-    ///
-    /// # Errors
-    /// Returns an error if broadcasting fails
-    pub async fn broadcast_area_event(
-        &self,
-        event_type: EventType,
-        _area_id: ThingsId,
-        data: Option<serde_json::Value>,
-        source: &str,
-    ) -> Result<()> {
-        let event = Event {
-            id: Uuid::new_v4(),
-            event_type,
-            timestamp: Utc::now(),
-            data,
-            source: source.to_string(),
-        };
-
-        self.broadcast(event).await
-    }
-
-    /// Create and broadcast a progress event
-    ///
-    /// # Errors
-    /// Returns an error if broadcasting fails
-    pub async fn broadcast_progress_event(
-        &self,
-        event_type: EventType,
-        _operation_id: Uuid,
-        data: Option<serde_json::Value>,
-        source: &str,
-    ) -> Result<()> {
-        let event = Event {
-            id: Uuid::new_v4(),
-            event_type,
-            timestamp: Utc::now(),
-            data,
-            source: source.to_string(),
-        };
-
-        self.broadcast(event).await
-    }
-
-    /// Convert a progress update to an event
-    ///
-    /// # Errors
-    /// Returns an error if broadcasting fails
-    pub async fn broadcast_progress_update(
-        &self,
-        update: ProgressUpdate,
-        source: &str,
-    ) -> Result<()> {
-        let event_type = match update.status {
-            crate::progress::ProgressStatus::Started => EventType::ProgressStarted {
-                operation_id: update.operation_id,
-            },
-            crate::progress::ProgressStatus::InProgress => EventType::ProgressUpdated {
-                operation_id: update.operation_id,
-            },
-            crate::progress::ProgressStatus::Completed => EventType::ProgressCompleted {
-                operation_id: update.operation_id,
-            },
-            crate::progress::ProgressStatus::Failed
-            | crate::progress::ProgressStatus::Cancelled => EventType::ProgressFailed {
-                operation_id: update.operation_id,
-            },
-        };
-
-        let data = serde_json::to_value(&update)?;
-        self.broadcast_progress_event(event_type, update.operation_id, Some(data), source)
-            .await
-    }
-
-    /// Get the number of active subscriptions
-    pub async fn subscription_count(&self) -> usize {
-        self.subscriptions.read().await.len()
-    }
-
-    /// Get a receiver for all events (unfiltered)
-    #[must_use]
-    pub fn subscribe_all(&self) -> broadcast::Receiver<Event> {
-        self.sender.subscribe()
-    }
-}
-
-impl Default for EventBroadcaster {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Event listener for handling events
-pub struct EventListener {
-    broadcaster: Arc<EventBroadcaster>,
-    #[allow(dead_code)]
-    subscriptions: Vec<Uuid>,
-}
-
-impl EventListener {
-    /// Create a new event listener
-    #[must_use]
-    pub fn new(broadcaster: Arc<EventBroadcaster>) -> Self {
-        Self {
-            broadcaster,
-            subscriptions: Vec::new(),
-        }
-    }
-
-    /// Subscribe to specific event types
-    pub async fn subscribe_to_events(
-        &mut self,
-        event_types: Vec<EventType>,
-    ) -> broadcast::Receiver<Event> {
-        let filter = EventFilter {
-            event_types: Some(event_types),
-            entity_ids: None,
-            sources: None,
-            since: None,
-        };
-
-        self.broadcaster.subscribe(filter).await
-    }
-
-    /// Subscribe to events for a specific entity
-    pub async fn subscribe_to_entity(&mut self, entity_id: ThingsId) -> broadcast::Receiver<Event> {
-        let filter = EventFilter {
-            event_types: None,
-            entity_ids: Some(vec![entity_id]),
-            sources: None,
-            since: None,
-        };
-
-        self.broadcaster.subscribe(filter).await
-    }
-
-    /// Subscribe to all events
-    #[must_use]
-    pub fn subscribe_to_all(&self) -> broadcast::Receiver<Event> {
-        self.broadcaster.subscribe_all()
-    }
-}
+pub use broadcaster::EventBroadcaster;
+pub use filter::{EventFilter, EventSubscription};
+pub use listener::EventListener;
+pub use types::{Event, EventType};
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use chrono::Utc;
+    use std::sync::Arc;
+    use things3_core::ThingsId;
+    use tokio::sync::broadcast;
+    use uuid::Uuid;
 
     #[test]
     fn test_event_creation() {
@@ -525,6 +134,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_progress_update_to_event() {
+        use crate::progress::ProgressUpdate;
         let broadcaster = EventBroadcaster::new();
         let mut receiver = broadcaster.subscribe_all();
 
@@ -830,8 +440,8 @@ mod tests {
     #[tokio::test]
     async fn test_event_listener_creation() {
         let broadcaster = EventBroadcaster::new();
-        let listener = EventListener::new(Arc::new(broadcaster));
-        assert_eq!(listener.subscriptions.len(), 0);
+        // Verify EventListener::new succeeds without panicking
+        let _listener = EventListener::new(Arc::new(broadcaster));
     }
 
     #[tokio::test]
@@ -1011,6 +621,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_event_broadcaster_broadcast_progress_update() {
+        use crate::progress::ProgressUpdate;
         let broadcaster = EventBroadcaster::new();
         let mut receiver = broadcaster.subscribe_all();
